// cabinet-administration/src/utils/enum.ts
export enum LocalStorageKeysEnum {
  TOKEN = 'token',
  ID = 'id',
  ROLE = 'role',
  NAME = 'name',
  PHONE = 'phone',
}

export const enum RequestStatusEnum {
  IDLE = 'idle',
  PENDING = 'pending',
  FULFILLED = 'fulfilled',
  REJECTED = 'rejected',
}

export enum HeaderIdEnum {
  ADMINISTRATION = '/administration',
  CREATE_CAKE = '/create-cake',
  CALENDAR = '/calendar',
  EDIT_CAKE = '/edit-cake',
  ANALYTICS = '/analytics',
  RECIPES = '/recipes',
  ORDERS = '/orders',
  ORDERS_HISTORY = '/history',
  TYPES = '/types',
  FEEDBACK = '/feedback',
}